requirements.txt
streamlit>=1.28.0
yfinance>=0.2.18
pandas>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
numpy>=1.24.0
